chatbot-model/requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
fastapi==0.66.0
gensim==4.0.1
html2text==2020.1.16
nest_asyncio==1.5.1
pandas==1.3.0
pyngrok==5.0.5
python-dotenv==0.18.0
spacy==3.1.0
summarizer==0.0.7
torch==1.9.0
transformers==4.8.2
uvicorn==0.14.0
